[TLX] Enable inlining in WS regions (#200)
Backporting upstream PR: https://github.com/triton-lang/triton/pull/6992
